backend/django/requirements.txt
asgiref==3.11.1
coverage==7.13.5
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
Pygments==2.20.0
PyJWT==2.12.1
PyMySQL==1.1.2
pytest==9.0.2
pytest-django==4.12.0
python-dotenv==1.2.2
setuptools==82.0.1
sqlparse==0.5.5
